// src/dorc-web/src/apis/dorc-api/models/EnvironmentContentBuildsApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface EnvironmentContentBuildsApiModel
 */
export interface EnvironmentContentBuildsApiModel {
    /**
     * @type {string}
     * @memberof EnvironmentContentBuildsApiModel
     */
    ComponentName?: string | null;
    /**
     * @type {string}
     * @memberof EnvironmentContentBuildsApiModel
     */
    RequestBuildNum?: string | null;
    /**
     * @type {number}
     * @memberof EnvironmentContentBuildsApiModel
     */
    RequestId?: number;
    /**
     * @type {string}
     * @memberof EnvironmentContentBuildsApiModel
     */
    UpdateDate?: string | null;
    /**
     * @type {string}
     * @memberof EnvironmentContentBuildsApiModel
     */
    State?: string | null;
}